Centralized the management of the application into dropsecure.py
dropsecure.py will provide the central interface for interacting
with the application. Current options include importing/exporting
of application keys, configuring the application as well as
starting and stopping the script.

To support starting and stopping the script, I have daemonized the
watcher.py file using the code sample from
http://www.jejik.com/articles/2007/02/a_simple_unix_linux_daemon_in_python/